Fix synchronization android test (and mark 2 of them as Flaky as real-api-dependent #MOV-507
