+++ synchronization/src/androidTest/java/de/cyface/synchronization/TestUtils.java
@@ -65,102 +65,4 @@ public final class TestUtils {
      */
     @SuppressWarnings("unused") // because this is used in the cyface flavour
     final static String TEST_API_URL = "https://s1.cyface.de:9090/api/v2";
-
-    /**
-     * Inserts a test {@link GeoLocation} into the database content provider accessed by the test.
-     *
-     * @param measurementIdentifier The identifier of the test {@link Measurement}.
-     * @param timestamp A fake test timestamp of the {@code GeoLocation}.
-     * @param lat The fake test latitude of the {@code GeoLocation}.
-     * @param lon The fake test longitude of the {@code GeoLocation}.
-     * @param speed The fake test speed of the {@code GeoLocation}.
-     * @param accuracy The fake test accuracy of the {@code GeoLocation}.
-     */
-    static void insertGeoLocation(final ContentResolver resolver, final String authority,
-            final long measurementIdentifier, final long timestamp, final double lat, final double lon,
-            final double speed, final int accuracy) {
-
-        ContentValues values = new ContentValues();
-        values.put(GeoLocationsTable.COLUMN_ACCURACY, accuracy);
-        values.put(GeoLocationsTable.COLUMN_GPS_TIME, timestamp);
-        values.put(GeoLocationsTable.COLUMN_LAT, lat);
-        values.put(GeoLocationsTable.COLUMN_LON, lon);
-        values.put(GeoLocationsTable.COLUMN_MEASUREMENT_FK, measurementIdentifier);
-        values.put(GeoLocationsTable.COLUMN_SPEED, speed);
-        resolver.insert(getGeoLocationsUri(authority), values);
-    }
-
-    /**
-     * Inserts a test {@code Measurement} into the database content provider accessed by the test. To add data to the
-     * {@code Measurement} use some or all of
-     * {@link #insertGeoLocation(ContentResolver, String, long, long, double, double, double, int)}
-     * {@link SharedTestUtils#insertTestAcceleration(Context, long, long, double, double, double)},
-     * {@link SharedTestUtils#insertTestRotation(Context, long, long, double, double, double)},
-     * {@link SharedTestUtils#insertTestDirection(Context, long, long, double, double, double)}
-     *
-     * @param vehicle The {@link Vehicle} type of the {@code Measurement}. A common value is {@link Vehicle#UNKNOWN} if
-     *            you do not care.
-     * @return The database identifier of the created {@link Measurement}.
-     */
-    static Measurement insertMeasurementEntry(final @NonNull PersistenceLayer persistence,
-            final @NonNull Vehicle vehicle) throws CursorIsNullException {
-
-        // usually called in DataCapturingService#Constructor
-        persistence.restoreOrCreateDeviceId();
-
-        return persistence.newMeasurement(vehicle);
-    }
-
-    /**
-     * This method inserts a {@link Measurement} into the persistence layer. Does not use the
-     * {@code CapturingPersistenceBehaviour} but the {@link DefaultPersistenceBehaviour}.
-     */
-    static Measurement insertSampleMeasurementWithData(@NonNull final Context context, final String authority,
-            final MeasurementStatus status, final PersistenceLayer persistence)
-            throws NoSuchMeasurementException, CursorIsNullException {
-
-        final Measurement measurement = insertMeasurementEntry(persistence, Vehicle.UNKNOWN);
-        final long measurementIdentifier = measurement.getIdentifier();
-        insertGeoLocation(context.getContentResolver(), authority, measurement.getIdentifier(), 1503055141000L,
-                49.9304133333333, 8.82831833333333, 0.0, 940);
-        insertPoint3d(context, measurementIdentifier, Point3dFile.ACCELERATIONS_FOLDER_NAME,
-                Point3dFile.ACCELERATIONS_FILE_EXTENSION, 1501662635973L, 10.1189575, -0.15088624, 0.2921924);
-        insertPoint3d(context, measurementIdentifier, Point3dFile.ROTATIONS_FOLDER_NAME,
-                Point3dFile.ROTATION_FILE_EXTENSION, 1501662635981L, 0.001524045, 0.0025423833, -0.0010279021);
-        insertPoint3d(context, measurementIdentifier, Point3dFile.DIRECTIONS_FOLDER_NAME,
-                Point3dFile.DIRECTION_FILE_EXTENSION, 1501662636010L, 7.65, -32.4, -71.4);
-
-        if (status == FINISHED || status == MeasurementStatus.SYNCED) {
-            // Store PointMetaData
-            final PointMetaData pointMetaData = new PointMetaData(1, 1, 1,
-                    MeasurementSerializer.PERSISTENCE_FILE_FORMAT_VERSION);
-            persistence.storePointMetaData(pointMetaData, measurementIdentifier);
-            persistence.setStatus(measurementIdentifier, FINISHED);
-        }
-
-        if (status == SYNCED) {
-            persistence.markAsSynchronized(measurement);
-        }
-
-        // Assert that data is in the database
-        final Measurement loadedMeasurement;
-        loadedMeasurement = persistence.loadMeasurement(measurementIdentifier);
-        assertThat(loadedMeasurement, notNullValue());
-        assertThat(persistence.loadMeasurementStatus(measurementIdentifier), is(equalTo(status)));
-
-        // Check the GeoLocations
-        List<GeoLocation> geoLocations = persistence.loadTrack(loadedMeasurement);
-        assertThat(geoLocations.size(), is(1));
-
-        // We can only check the PointMetaData for measurements which are not open anymore (else it's still in cache)
-        if (status != OPEN) {
-            final PointMetaData pointMetaData = persistence.loadPointMetaData(measurementIdentifier);
-            assertThat(pointMetaData.getAccelerationPointCounter(), is(equalTo(1)));
-            assertThat(pointMetaData.getRotationPointCounter(), is(equalTo(1)));
-            assertThat(pointMetaData.getDirectionPointCounter(), is(equalTo(1)));
-            assertThat(pointMetaData.getPersistenceFileFormatVersion(),
-                    is(equalTo(MeasurementSerializer.PERSISTENCE_FILE_FORMAT_VERSION)));
-        }
-        return measurement;
-    }
 }
