+++ testutils/src/main/java/de/cyface/testutils/SharedTestUtils.java
@@ -4,7 +4,14 @@ import static de.cyface.persistence.Constants.TAG;
 import static de.cyface.persistence.Utils.getGeoLocationsUri;
 import static de.cyface.persistence.Utils.getIdentifierUri;
 import static de.cyface.persistence.Utils.getMeasurementUri;
+import static de.cyface.persistence.model.MeasurementStatus.FINISHED;
+import static de.cyface.persistence.model.MeasurementStatus.OPEN;
+import static de.cyface.persistence.model.MeasurementStatus.SYNCED;
 import static de.cyface.persistence.serialization.MeasurementSerializer.BYTES_IN_ONE_POINT_3D_ENTRY;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.junit.Assert.assertThat;
 
 import java.io.File;
 import java.nio.ByteBuffer;
@@ -13,16 +20,26 @@ import java.util.ArrayList;
 import java.util.List;
 
 import android.content.ContentResolver;
+import android.content.ContentValues;
 import android.content.Context;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import de.cyface.persistence.DefaultFileAccess;
+import de.cyface.persistence.DefaultPersistenceBehaviour;
 import de.cyface.persistence.FileAccessLayer;
+import de.cyface.persistence.GeoLocationsTable;
 import de.cyface.persistence.IdentifierTable;
+import de.cyface.persistence.NoSuchMeasurementException;
+import de.cyface.persistence.PersistenceLayer;
 import de.cyface.persistence.model.GeoLocation;
 import de.cyface.persistence.model.Measurement;
+import de.cyface.persistence.model.MeasurementStatus;
 import de.cyface.persistence.model.Point3d;
+import de.cyface.persistence.model.PointMetaData;
+import de.cyface.persistence.model.Vehicle;
+import de.cyface.persistence.serialization.MeasurementSerializer;
 import de.cyface.persistence.serialization.Point3dFile;
+import de.cyface.utils.CursorIsNullException;
 import de.cyface.utils.Validate;
 
 /**
@@ -36,72 +53,19 @@ import de.cyface.utils.Validate;
 public class SharedTestUtils {
 
     /**
-     * Inserts a test acceleration into the database content provider accessed by the test.
-     *
-     * @param context The {@link Context} required to access the persistence layer
-     * @param measurementId The id of the test {@link Measurement}.
-     * @param timestamp A fake test timestamp of the acceleration.
-     * @param x A fake test x coordinate of the acceleration.
-     * @param y A fake test y coordinate of the acceleration.
-     * @param z A fake test z coordinate of the acceleration.
-     */
-    public static void insertTestAcceleration(@NonNull final Context context, final long measurementId,
-            final long timestamp, final double x, final double y, final double z) {
-        insertPoint3d(context, measurementId, Point3dFile.ACCELERATIONS_FOLDER_NAME,
-                Point3dFile.ACCELERATIONS_FILE_EXTENSION, timestamp, x, y, z);
-    }
-
-    /**
-     * Inserts a test rotation into the database content provider accessed by the test.
-     *
-     * @param context The {@link Context} required to access the persistence layer
-     * @param measurementId The id of the test {@link Measurement}.
-     * @param timestamp A fake test timestamp of the direction.
-     * @param x A fake test x coordinate of the direction.
-     * @param y A fake test y coordinate of the direction.
-     * @param z A fake test z coordinate of the direction.
-     */
-    public static void insertTestRotation(@NonNull final Context context, final long measurementId,
-            final long timestamp, final double x, final double y, final double z) {
-        insertPoint3d(context, measurementId, Point3dFile.ROTATIONS_FOLDER_NAME, Point3dFile.ROTATION_FILE_EXTENSION,
-                timestamp, x, y, z);
-    }
-
-    /**
-     * Inserts a test direction into the database content provider accessed by the test.
-     *
-     * @param context The {@link Context} required to access the persistence layer
-     * @param measurementId The id of the test {@link Measurement}.
-     * @param timestamp A fake test timestamp of the direction.
-     * @param x A fake test x coordinate of the direction.
-     * @param y A fake test y coordinate of the direction.
-     * @param z A fake test z coordinate of the direction.
-     */
-    public static void insertTestDirection(@NonNull final Context context, final long measurementId,
-            final long timestamp, final double x, final double y, final double z) {
-        insertPoint3d(context, measurementId, Point3dFile.DIRECTIONS_FOLDER_NAME, Point3dFile.DIRECTION_FILE_EXTENSION,
-                timestamp, x, y, z);
-    }
-
-    /**
      * Inserts a test {@link Point3d} into the database content provider accessed by the test.
      *
-     * @param context The {@link Context} required to access the persistence layer
-     * @param measurementId The id of the test {@link Measurement}.
-     * @param folderName The folder name defining the {@link Point3d} type of the file
-     * @param fileExtension the extension of the file type
+     * @param point3dFile existing file to append the data to
      * @param timestamp A fake test timestamp of the {@code Point3d}.
      * @param x A fake test x coordinate of the {@code Point3d}.
      * @param y A fake test y coordinate of the {@code Point3d}.
      * @param z A fake test z coordinate of the {@code Point3d}.
      */
-    public static void insertPoint3d(@NonNull final Context context, final long measurementId,
-            @NonNull final String folderName, @NonNull final String fileExtension, final long timestamp, final double x,
+    public static void insertPoint3d(@NonNull final Point3dFile point3dFile, final long timestamp, final double x,
             final double y, final double z) {
-        final Point3dFile file = new Point3dFile(context, measurementId, folderName, fileExtension);
         final List<Point3d> points = new ArrayList<>();
         points.add(new Point3d((float)x, (float)y, (float)z, timestamp));
-        file.append(points);
+        point3dFile.append(points);
     }
 
     /**
@@ -210,5 +174,107 @@ public class SharedTestUtils {
         removedDatabaseRows += removedGeoLocations;
         removedDatabaseRows += removedMeasurements;
         return removedFiles + removedDatabaseRows;
+    }
+
+    /**
+     * This method inserts a {@link Measurement} into the persistence layer. Does not use the
+     * {@code CapturingPersistenceBehaviour} but the {@link DefaultPersistenceBehaviour}.
+     */
+    public static Measurement insertSampleMeasurementWithData(@NonNull final Context context, final String authority,
+            final MeasurementStatus status, final PersistenceLayer persistence)
+            throws NoSuchMeasurementException, CursorIsNullException {
+
+        final Measurement measurement = insertMeasurementEntry(persistence, Vehicle.UNKNOWN);
+        final long measurementIdentifier = measurement.getIdentifier();
+        insertGeoLocation(context.getContentResolver(), authority, measurement.getIdentifier(), 1503055141000L,
+                49.9304133333333, 8.82831833333333, 0.0, 940);
+
+        // Insert file base data
+        final Point3dFile accelerationsFile = new Point3dFile(context, measurementIdentifier,
+                Point3dFile.ACCELERATIONS_FOLDER_NAME, Point3dFile.ACCELERATIONS_FILE_EXTENSION);
+        final Point3dFile rotationsFile = new Point3dFile(context, measurementIdentifier,
+                Point3dFile.ROTATIONS_FOLDER_NAME, Point3dFile.ROTATION_FILE_EXTENSION);
+        final Point3dFile directionsFile = new Point3dFile(context, measurementIdentifier,
+                Point3dFile.DIRECTIONS_FOLDER_NAME, Point3dFile.ROTATION_FILE_EXTENSION);
+        insertPoint3d(accelerationsFile, 1501662635973L, 10.1189575, -0.15088624, 0.2921924);
+        insertPoint3d(rotationsFile, 1501662635981L, 0.001524045, 0.0025423833, -0.0010279021);
+        insertPoint3d(directionsFile, 1501662636010L, 7.65, -32.4, -71.4);
+
+        if (status == FINISHED || status == MeasurementStatus.SYNCED) {
+            // Store PointMetaData
+            final PointMetaData pointMetaData = new PointMetaData(1, 1, 1,
+                    MeasurementSerializer.PERSISTENCE_FILE_FORMAT_VERSION);
+            persistence.storePointMetaData(pointMetaData, measurementIdentifier);
+            persistence.setStatus(measurementIdentifier, FINISHED);
+        }
+
+        if (status == SYNCED) {
+            persistence.markAsSynchronized(measurement);
+        }
+
+        // Assert that data is in the database
+        final Measurement loadedMeasurement;
+        loadedMeasurement = persistence.loadMeasurement(measurementIdentifier);
+        assertThat(loadedMeasurement, notNullValue());
+        assertThat(persistence.loadMeasurementStatus(measurementIdentifier), is(equalTo(status)));
+
+        // Check the GeoLocations
+        List<GeoLocation> geoLocations = persistence.loadTrack(loadedMeasurement);
+        assertThat(geoLocations.size(), is(1));
+
+        // We can only check the PointMetaData for measurements which are not open anymore (else it's still in cache)
+        if (status != OPEN) {
+            final PointMetaData pointMetaData = persistence.loadPointMetaData(measurementIdentifier);
+            assertThat(pointMetaData.getAccelerationPointCounter(), is(equalTo(1)));
+            assertThat(pointMetaData.getRotationPointCounter(), is(equalTo(1)));
+            assertThat(pointMetaData.getDirectionPointCounter(), is(equalTo(1)));
+            assertThat(pointMetaData.getPersistenceFileFormatVersion(),
+                    is(equalTo(MeasurementSerializer.PERSISTENCE_FILE_FORMAT_VERSION)));
+        }
+        return measurement;
+    }
+
+    /**
+     * Inserts a test {@code Measurement} into the database content provider accessed by the test. To add data to the
+     * {@code Measurement} use some or all of
+     * {@link #insertGeoLocation(ContentResolver, String, long, long, double, double, double, int)}
+     * {@link SharedTestUtils#insertPoint3d(Point3dFile, long, double, double, double)} (Context, long, long, double,
+     * double, double)},
+     *
+     * @param vehicle The {@link Vehicle} type of the {@code Measurement}. A common value is {@link Vehicle#UNKNOWN} if
+     *            you do not care.
+     * @return The database identifier of the created {@link Measurement}.
+     */
+    public static Measurement insertMeasurementEntry(final @NonNull PersistenceLayer persistence,
+            final @NonNull Vehicle vehicle) throws CursorIsNullException {
+
+        // usually called in DataCapturingService#Constructor
+        persistence.restoreOrCreateDeviceId();
+
+        return persistence.newMeasurement(vehicle);
+    }
+
+    /**
+     * Inserts a test {@link GeoLocation} into the database content provider accessed by the test.
+     *
+     * @param measurementIdentifier The identifier of the test {@link Measurement}.
+     * @param timestamp A fake test timestamp of the {@code GeoLocation}.
+     * @param lat The fake test latitude of the {@code GeoLocation}.
+     * @param lon The fake test longitude of the {@code GeoLocation}.
+     * @param speed The fake test speed of the {@code GeoLocation}.
+     * @param accuracy The fake test accuracy of the {@code GeoLocation}.
+     */
+    public static void insertGeoLocation(final ContentResolver resolver, final String authority,
+            final long measurementIdentifier, final long timestamp, final double lat, final double lon,
+            final double speed, final int accuracy) {
+
+        ContentValues values = new ContentValues();
+        values.put(GeoLocationsTable.COLUMN_ACCURACY, accuracy);
+        values.put(GeoLocationsTable.COLUMN_GPS_TIME, timestamp);
+        values.put(GeoLocationsTable.COLUMN_LAT, lat);
+        values.put(GeoLocationsTable.COLUMN_LON, lon);
+        values.put(GeoLocationsTable.COLUMN_MEASUREMENT_FK, measurementIdentifier);
+        values.put(GeoLocationsTable.COLUMN_SPEED, speed);
+        resolver.insert(getGeoLocationsUri(authority), values);
     }
 }
